+++ src/index.js
@@ -3,6 +3,9 @@ import './init';
 export {VERSION} from './lib/utils/globals';
 
 export {default as Log} from './lib/log';
+export {COLOR} from './lib/utils/color';
+
+// export {default as Stats} from './lib/stats';
 
 // Experimental exports
 import {enableDOMLogging} from './lib/utils/log-to-dom';