+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 
-v1.0.0-alpha.2
+v1.0.0-alpha.4
+- Support saving probe config in local storage
+- Doc improvements
+
+v1.0.0-alpha.3
 - Fix groupEnd issue under Node.js
 
 v1.0.0-alpha.2